feat: elevated utils and types to app level
