+++ frontend/src/components/lint-results.tsx
