+++ frontend/src/components/pages/rp-connect/Pipelines.List.tsx
@@ -40,7 +40,10 @@ import type { AddTopicFormData, AddUserFormData, ConnectTilesFormData } from './
 
 const { ToastContainer, toast } = createStandaloneToast();
 
-const CreatePipelineButton = () => {
+/**
+ * Navigates to /rp-connect/create (legacy flow)
+ */
+const LegacyCreatePipelineButton = () => {
   return (
     <div>
       <NewButton as={Link} to="/rp-connect/create">
@@ -50,7 +53,10 @@ const CreatePipelineButton = () => {
   );
 };
 
-const NewCreatePipelineButton = () => {
+/**
+ * Navigates to wizard and clears session storage
+ */
+const WizardCreatePipelineButton = () => {
   const [_, setPersistedConnectionName] = useSessionStorage<Partial<ConnectTilesFormData>>(
     CONNECT_WIZARD_CONNECTOR_KEY,
     {},
@@ -73,7 +79,23 @@ const NewCreatePipelineButton = () => {
   );
 };
 
-const EmptyPlaceholder = () => {
+/**
+ * Shows image, text, and create button
+ */
+const LegacyEmptyState = () => {
+  return (
+    <Flex alignItems="center" justifyContent="center" flexDirection="column" gap="4" mb="4">
+      <Image src={EmptyConnectors} />
+      <Box>You have no Redpanda Connect pipelines.</Box>
+      <LegacyCreatePipelineButton />
+    </Flex>
+  );
+};
+
+/**
+ * Shows ConnectTiles and navigates to wizard with connector pre-selected
+ */
+const WizardEmptyState = () => {
   const [, setPersistedConnectionName] = useSessionStorage<Partial<ConnectTilesFormData>>(
     CONNECT_WIZARD_CONNECTOR_KEY,
     {},
@@ -81,7 +103,6 @@ const EmptyPlaceholder = () => {
   const [, setPersistedTopic] = useSessionStorage<Partial<AddTopicFormData>>(CONNECT_WIZARD_TOPIC_KEY, {});
   const [, setPersistedUser] = useSessionStorage<Partial<AddUserFormData>>(CONNECT_WIZARD_USER_KEY, {});
   const navigate = useNavigate();
-  const enableConnectTiles = isFeatureFlagEnabled('enableRpcnTiles');
 
   const handleConnectionChange = useCallback(
     (connectionName: string, connectionType: ConnectComponentType) => {
@@ -103,19 +124,12 @@ const EmptyPlaceholder = () => {
     [setPersistedConnectionName, setPersistedTopic, setPersistedUser, navigate],
   );
 
-  const oldUI = (
-    <Flex alignItems="center" justifyContent="center" flexDirection="column" gap="4" mb="4">
-      <Image src={EmptyConnectors} />
-      <Box>You have no Redpanda Connect pipelines.</Box>
-      <CreatePipelineButton />
-    </Flex>
-  );
+  return <ConnectTiles onChange={handleConnectionChange} componentTypeFilter={['input', 'output']} hideHeader />;
+};
 
-  return enableConnectTiles ? (
-    <ConnectTiles onChange={handleConnectionChange} componentTypeFilter={['input', 'output']} hideHeader />
-  ) : (
-    oldUI
-  );
+const EmptyPlaceholder = () => {
+  const enableRpcnTiles = isFeatureFlagEnabled('enableRpcnTiles');
+  return enableRpcnTiles ? <WizardEmptyState /> : <LegacyEmptyState />;
 };
 
 export const PipelineStatus = observer((p: { status: Pipeline_State }) => {
@@ -246,11 +260,11 @@ class RpConnectPipelinesList extends PageComponent<{}> {
 
         {pipelinesApi.pipelines.length !== 0 && isFeatureFlagEnabled('enableRpcnTiles') ? (
           <div className="my-5">
-            <NewCreatePipelineButton />
+            <WizardCreatePipelineButton />
           </div>
         ) : pipelinesApi.pipelines.length !== 0 ? (
           <div className="flex flex-col gap-2 my-5">
-            <CreatePipelineButton />
+            <LegacyCreatePipelineButton />
             <SearchField
               width="350px"
               searchText={uiSettings.pipelinesList.quickSearch}
